% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar_cjsg_trf1.R
\name{baixar_cjsg_trf1}
\alias{baixar_cjsg_trf1}
\title{Baixa decisões de segunda instância do TRF1}
\usage{
baixar_cjsg_trf1(
  livre = "",
  aspas = FALSE,
  tipo = "ACORDAO",
  data_inicial = "",
  data_final = "",
  diretorio = "."
)
}
\arguments{
\item{livre}{busca livre}

\item{aspas}{TRUE para colocar a busca entre aspas}

\item{tipo}{Default para "ACÓRDÃOS"}

\item{data_inicial}{formato "dd/mm/aaaa"}

\item{data_final}{formato "dd/mm/aaaa"}

\item{diretorio}{Default para atual.}
}
\value{
Baixa os htmls nos diretório especificado
}
\description{
Baixa decisões de segunda instância do TRF1
}
\examples{
\dontrun{
baixar_cjsg_trf1(
  livre = "", tipo = "ACORDAO",
  data_inicial = "10/07/2019", data_final = "31/07/2019"
)
}

}
