% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar_docs_trf1.R
\name{baixar_docs_trf1}
\alias{baixar_docs_trf1}
\title{Baixar documentos listados em extrair_tabela_docs_trf1}
\usage{
baixar_docs_trf1(df = NULL, diretorio = ".", tudo = FALSE)
}
\arguments{
\item{df}{tibble lido com ler_por_data_trf1}

\item{diretorio}{diretório}

\item{tudo}{default para FALSE, ou seja, mantêm somente ementas, decisões,
relatórios     e votos}
}
\value{
docs
}
\description{
Baixar documentos listados em extrair_tabela_docs_trf1
}
\examples{
\dontrun{
baixar_docs_trf1(df)
}
}
