% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar_por_data_trf1.R
\name{baixar_por_data_trf1}
\alias{baixar_por_data_trf1}
\title{Baixar lista de decisões do TRF1 por data}
\usage{
baixar_por_data_trf1(data = NULL, diretorio = ".")
}
\arguments{
\item{data}{data da decisão. Data de hoje, se não informado.}

\item{diretorio}{diretório, atual se não informado.}
}
\value{
html
}
\description{
Baixar lista de decisões do TRF1 por data
}
\examples{
\dontrun{
baixar_por_data_trf1()
}
}
