% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar_processo_trf1.R
\name{baixar_processo_trf1}
\alias{baixar_processo_trf1}
\title{Baixar processos do trf1 pelo número}
\usage{
baixar_processo_trf1(x, diretorio = ".")
}
\arguments{
\item{x}{vetor com os números dos processos}

\item{diretorio}{diretório}
}
\value{
htmls no diretório indicado
}
\description{
Baixar processos do trf1 pelo número
}
\examples{
\dontrun{
baixar_processo_trf1(x="00010419820184013300",diretorio = ".")
}
}
