% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar_url_detalhes_trf1.R
\name{baixar_url_detalhes_trf1}
\alias{baixar_url_detalhes_trf1}
\title{Baixar detalhes dos processos lidos com ler_por_data_trf1}
\usage{
baixar_url_detalhes_trf1(df = NULL, diretorio = ".")
}
\arguments{
\item{df}{tibble lido com ler_por_data_trf1}

\item{diretorio}{diretório}
}
\value{
html
}
\description{
Baixar detalhes dos processos lidos com ler_por_data_trf1
}
\examples{
\dontrun{
baixar_url_detalhes_trf1(df)
}
}
