% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codigo_trf1.R
\docType{data}
\name{codigo_trf1}
\alias{codigo_trf1}
\title{Códigos das unidades judiciárias do  TRF1}
\format{
tabela com 98 linhas e 4 colunas:
\describe{
  \item{\code{unidade}}{Unidade judiciária}
  \item{\code{sigla}}{Sigla da unidade judiciária}
  \item{\code{codigo}}{Código da unidade judiciária}
  \item{\code{turma_codigo}}{Código da Turma}
}
}
\source{
\url{https://trf1.jus.br/}
}
\usage{
codigo_trf1
}
\description{
Códigos dos distribuidores do TRF1
}
\keyword{datasets}
