% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrair_docs_por_processo.R
\name{extrair_docs_por_processo}
\alias{extrair_docs_por_processo}
\title{Extrai tabela urls dos documentos com base no número do processo}
\usage{
extrair_docs_por_processo(processos = NULL)
}
\arguments{
\item{processos}{vetor com números dos processos}
}
\value{
tibble com processo,documento, data,url_documento e extensão
}
\description{
Extrai tabela urls dos documentos com base no número do processo
}
