% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrair_relator.R
\name{extrair_relator}
\alias{extrair_relator}
\title{Extrai nome do relator dos textos dos julgados}
\usage{
extrair_relator(julgado = NULL)
}
\arguments{
\item{julgado}{Texto do julgado}
}
\value{
Vetor com os nomes dos relatores, sem acento e em minúsculo
}
\description{
Extrai nome do relator dos textos dos julgados
}
