% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrair_tabela_docs_trf1.R
\name{extrair_tabela_docs_trf1}
\alias{extrair_tabela_docs_trf1}
\title{Extrai urls dos documentos a partir da tabela criada com ler_por_data}
\usage{
extrair_tabela_docs_trf1(df = NULL)
}
\arguments{
\item{df}{tibble com número do processo e url_inteiro_teor}
}
\value{
nova tabela
}
\description{
Extrai urls dos documentos a partir da tabela criada com ler_por_data
}
\examples{
\dontrun{
df_docs <- extrair_tabela_docs_trf1(df)
}
}
