% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_detalhes_trf1.R
\name{ler_detalhes_trf1}
\alias{ler_detalhes_trf1}
\title{ler detalhes baixados com baixar_url_detalhes_trf1}
\usage{
ler_detalhes_trf1(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{vetor caminhos para arquivos}

\item{diretorio}{se arquivos for NULL, informar diretorio}
}
\value{
tibble
}
\description{
ler detalhes baixados com baixar_url_detalhes_trf1
}
\examples{
\dontrun{
df <- ler_detalhes_stj()
}
}
