% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_docs_trf1.R
\name{ler_docs_trf1}
\alias{ler_docs_trf1}
\title{Lê lodocumentos decisórios}
\usage{
ler_docs_trf1(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Se não informados, informar o diretório
onde se encontram.}

\item{diretorio}{Informar se não informou arquivos}
}
\value{
tibble
}
\description{
Lê lodocumentos decisórios
}
\examples{
\dontrun{
df <- ler_docs_trf1()
}
}
