% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_partes_trf1.R
\name{ler_partes_trf1}
\alias{ler_partes_trf1}
\title{ler partes do TRF1}
\usage{
ler_partes_trf1(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{diretorio}{Informar se vetor de arquivos for NULL}
}
\value{
tibble
}
\description{
ler partes do TRF1
}
\examples{
\dontrun{
ler_partes_trf1()
}
}
