% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_por_data_trf1.R
\name{ler_por_data_trf1}
\alias{ler_por_data_trf1}
\title{Ler arquivos baixados por baixar_por_data_trf1}
\usage{
ler_por_data_trf1(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Se não informados, buscará diretório}

\item{diretorio}{Se arquivos não forem informados, buscará
no diretório indicado}
}
\value{
data.frame com urls dos andamentos e dos inteiros teores
}
\description{
Ler arquivos baixados por baixar_por_data_trf1
}
\examples{
\dontrun{
df <- ler_por_data_trf1()
}
}
