% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trf1_baixar_cpopg.R
\name{trf1_baixar_cpopg}
\alias{trf1_baixar_cpopg}
\title{Baixar processos de primeira instância}
\usage{
trf1_baixar_cpopg(processo = NULL, secao = NULL, diretorio = ".")
}
\arguments{
\item{processo}{Vetor com número dos processos}

\item{secao}{Seção judiciária}

\item{diretorio}{Diretório}
}
\value{
Htmls
}
\description{
Baixar processos de primeira instância
}
