% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trf1_baixar_cposg.R
\name{trf1_baixar_cposg}
\alias{trf1_baixar_cposg}
\title{Baixar processos da segunda instância do TRF1}
\usage{
trf1_baixar_cposg(processo = NULL, diretorio = ".", partes = TRUE)
}
\arguments{
\item{processo}{Número do processo}

\item{diretorio}{Diretório}

\item{partes}{A partes são baixadas num arquivo separado}
}
\value{
Dois htmls, um com dados e andamentos processuis, outro com as partes.
}
\description{
Baixar processos da segunda instância do TRF1
}
