% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trf1_ler_cpopg_docs.R
\name{trf1_ler_cpopg_docs}
\alias{trf1_ler_cpopg_docs}
\title{Lê julgados de primeiro grau em texto}
\usage{
trf1_ler_cpopg_docs(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{diretorio}{Informar se vetor de arquivos for NULL}
}
\value{
tibble
}
\description{
Lê julgados de primeiro grau em texto
}
