% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trf1_ler_dados_cpopg.R
\name{trf1_ler_dados_cpopg}
\alias{trf1_ler_dados_cpopg}
\title{Lê dados da consulta de processos de primeiro grau}
\usage{
trf1_ler_dados_cpopg(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{diretorio}{Informar se não fornecer arquivos}
}
\value{
tibble
}
\description{
Lê dados da consulta de processos de primeiro grau
}
