% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trf1_ler_julgados_imagens.R
\name{trf1_ler_julgados_imagens}
\alias{trf1_ler_julgados_imagens}
\title{Lê sentenças de primeiro grau de pdf em imagens}
\usage{
trf1_ler_julgados_imagens(arquivos = NULL)
}
\arguments{
\item{arquivos}{Vetor de arquivos}
}
\value{
Tibble
}
\description{
Lê sentenças de primeiro grau de pdf em imagens
}
\details{
O datafrane criado com a trf1_ler_cpopg_docs
     informa se é imagem ou não.
}
