% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trf1_quantidade.R
\name{trf1_quantidade}
\alias{trf1_quantidade}
\title{Calcula quantidade de processos distribuídos por fórum}
\usage{
trf1_quantidade(
  inicio = 0,
  fim = NULL,
  ano = NULL,
  distribuidor = NULL,
  secao = NULL,
  diretorio = "data-raw"
)
}
\arguments{
\item{inicio}{Geralmente 1}

\item{fim}{Inteiro grande o suficiente}

\item{ano}{Ano}

\item{distribuidor}{Verifique os códigos em codigos2}

\item{secao}{sigla da seção}

\item{diretorio}{Um diretório vazio para downloads provisórios}

\item{funcao}{Geralmente baixar_cposg}
}
\value{
vetor
}
\description{
Calcula quantidade de processos distribuídos por fórum
}
